Make game build again.
Note that there is some work that needs to be done in order to properly
decouple between engine and game. Things are pretty mixed up there, but
this will be done with time in a separate PR, as it requires some
thinking before doing due to the UI's implementation that maybe deserve
a better design in terms of abstraction, so that game and engine are not
tightly coupled.