Disable Ebenezer optimizations in release mode.
Don't ask why...compiler explodes on CMagicProcess::MagicPacket where it
excutes all types of magic. I think it gets confused by the amount of
branches this function produces when compiling and when removing the
group of switch cases there, it compiles.

So temporarily we disable optimizations, even though in Release mode and
it will be fixed once we do refactoring to this implementation.