Fix outer loop access instances.
I used the following regex pattern to find them:
(for\s*\(\s*(i|j)\s*=)|(\s*int\s*(i|j)\s*;)

Note that this was done with extra cautious, because there were some instances
where it was obviously necessary for the variable to be outside if it's
being accessed in another usage.
Other instances were not modified, as they flow of the function with
goto functions was too hectic to change. Mostly noticed through the
server codebase, where the engineers coded as if we're doing C in C++.