Fix bad usages of std::vector buffers.
Note that there is no need to use std::vector<char> as buffers when the
size is known as a constant value.
If the size is known, use standard arrays to reduce waste of resources.
This change also fixes an issue where some buffers had size + 1, because
we use begin and end, there is no need to account for the null
termination, since the std::string constructor will already add it.