Rewrite re version of v1264 leveling guide.
Note that the previous implementation was missing a lot of details, as
well as not using the correct data structures to make the resource
allocation more dynamic.

Now that everything is reversed engineered with this uif, we can
gurantee that it works exactly the same as per official v1264 client.