Implement GetStringAsInt convenient function.
This in particular will be useful when we formatting coins and then
retrieving them back from the UI before casting to integer.
I added it into CN3UIString, since it's implemented in such a way that
it may be used for other scenarios where the string contain more
characters that should be deleted before attempting the cast to in
integer.