Fix when exiting game to not show elapsed message.
As per code-review comment that I had, when using PostQuitMessage, the
code below can be executed since it's a scheduled exit.
This commit will fix the issue when we exit or select character, the
elapsed red message in the chat won't show up unless we were really
attacked or attacking.