Test mode toggle fix
Test mode now defaults to 0
Add the following to the aiserver.ini
[TEST]
MODE=0
Test mode will make your R hit 3000+ dmg  and make mobs hit you for 10 dmg, this makes debugging debuff skills harder on another player when a mob attacks them. Since we are not at that stage yet we should still do this because it makes life easier.
https://imgur.com/a/5O5cQhm