Fixed targetbar HP updating
Since we don't have states yet for debuffs stuns slows burns I decided to make them update as HP decreases. I did not want to hide the other progressbar states, the issue is with the other progressbars overlapping the red one, so even though it looks like it is not moving it is but we just don't perceive it due to the red bar being at the lowest order. We can hide the other ones by simply doing:
        m_pProgressHP_burn->SetVisible(false);
        m_pProgressHP_curse->SetVisible(false);
        m_pProgressHP_slow->SetVisible(false);