OPD to N3ME
Ingame OPD object parsing and exporting to N3ME sdt (object) file extension. Ugly code is not always the best, but it should be easy to understand. It reads shape count from OPD (ISC) then creates a filestream named mapexported.sdt in main client directory. Then it prints(echo) it to console. Once it reaches pshape->load(hfile) it gets the pos,scale,rot and saves it in a vector3 (x.y.z). It saves the partcount to ISPC as an integer. Remove the 0 from the string and so we can append .n3shape. Writes it to the file Then it does a diffuse ARGB for each pshape part it comes across. At the end of the load handle it flushes the stream so it saves and can be opened by notepad. This is all self explanatory but I decided to make a detailed repo for those that don't understand the code.
-Tahsin