1264 Terrain GTD loading
This ugly code loads 1264 maps, it reads the newly added information which are:
1 version
2 map name length
3 actual name of the map.
If the version is not 1 then resets the file read pointer to the beginning of the file and continues like nothing happened. The string is not null terminated so when Steve is back I hope this can be cleaned up but for now it suffices. Try to intermix maps (GTD) in your zone directory and let me know. Those could be 1097 or 1264 myko maps. I tested 2 maps they worked, so feedback is welcome!