UISkillTreeDlg Mana episode 1
Mana was not showing for type 1 and type 2 attacks (melee+archer). This is because the code was designed to calculate mana based on how much damage you are doing to the mob. This is obnoxious to me to the point that I am having diarrhea issues as all my nutrients are drained, and no mana potions seem sufficient enough to replenish it. Jokes aside all skills should show the amount of mana being drained no matter if they are type 1 (melee) or buffs/heals!