1264 Loading Magic tbls and magic db issues
If mobs are allowed to have over 255 resistances then so should buffs/skills and users be able to. This comes with current DB not accepting negative values in the database for tinyint. Some debuffs give negative numerical values. We cant do % here because something that gives 5 str gives 5 str and something that takes str does -5 str so % is not an option.