Fix CN3FXPartParticles::Duplicate bug with particles gather point.
This fixes the GatherPoint for particles to return to a specific spot,
there are3 types spread gather and none, idk if none is ever used but
I do sure know staticnova,1920 heal and some other skills use a
GatherPoint. This will fix the Particles from collecting in the terrain
instead of towards the char's head (1920 heal).