Fix invalid SQLExecDirect() calls These should supply the length of the query or SQL_NTS to indicate a null-terminated string, not the entire size of the buffer.
This only happens to behave 'correctly' in older ANSI drivers by luck; in most cases, modern drivers will
report an error with the query which includes the trailing \0 bytes from the buffer.

For simplicity's sake, we can just supply SQL_NTS here and let the driver handle detection.
This can be improved later to directly specify the actual length of the string when it's already known,
which wsprintf() returns (and is used by most if not all instances).