Append DirectX runtime dlls to user's PATH during debugging.
Note that most users don't have Microsoft DirectX 9 2010 June installed
widely on their system, which means they will miss the runtime dlls for
DirectX math library. Therefore in case they didn't install it, we
append them to the PATH environment variable for the debugger, so that
it will pickup from the portable SDKs if not found in system.