Update Zip compression project to link against zlib.
Note that in the previous 1.0.0 ko-vendor build of zlib, we provided pdb
for symbols, which LoginServer wouldn't be able to find, hence it was
linking against zlib as well. However with the 1.1.0 ko-vendor build of
zlib, we now embedded the symbols into it with /Z7 flag, so that the
symbols are attached into each obj file inside the static library.
Therefore there is no longer need for a pdb and hence LoginServer now
only needs to link against Zip and Zip against zlib.