Update all resource files (*.rc) via latest resource editor.
Note that newer versions of resource editors bundled in Visual Studio
formats these files differently and also deprecates some old usages.
Therefore I made this round so that if someone edit them in the future,
the diff won't be bloated as much and it will be easier to review the
changes.