Add back text=auto to gitattributes.
Note that I had issues with that where I missed one file to specify as binary,
which is *.cur files and then the resource compiler had issues to compile the
resource, since git normalized it assuming it is a text file.
Using text=auto we'll let git detect if it is a text file and only then apply
the rule for line feed as needed.