Ensure correct handling of null-termination in N3ME string writing.
The `+1` adjustment ensures that the null-termination byte is included
when writing to file, resolving a previous issue where the algorithm
incorrectly calculated the string length while performing a reverse loop
to extract the filename from a path.

This issue, which persists in official maps from the 2xxx versions, may
lead to confusion for those interpreting these formats. The root cause
is that, when reading strings from files, 4 bytes are consistently read
for the string length, followed by the string data based on that length.
If the null-termination byte is read directly into a dynamically
allocated container, it will be included, resulting in failed string
comparisons even though the visible string values appear identical.

To maintain compatibility with official formats, we must include the
null-termination byte.

The issue remained unnoticed because the client reads the string into a
`char buffer[len]` and then constructs an `std::string` from it. While
the redundant null byte is present in the binary format, `std::string`
automatically constructs from `char buffer` up to the null-termination.
Thus, the binary format's string length is effectively 1 byte longer
than necessary, making this a hidden bug.