Remove unused grass (seed) implementation from both client and N3ME.
In N3ME, this feature is referred to as "seed," while in the client, it
is called "grass". This functionality is unused in the official release
and does not justify ongoing maintenance.

Reviewing the code and files in Misc/grass, which haven't been modified
since the game's release, confirms it is unused. The original developer
appears to have intended for grass/seeds to be part of terrain data (gtd).
However, in the official implementation, they are added to the opd file
as shapes with simple textures and without collision, making the grass
implementation redundant.