Remove unused branch tags from .gitmodules file.
The `branch` option in `.gitmodules` is only relevant when tracking
remote changes to a specific branch, such as a dedicated development
branch. However, we use `master` as both the default development branch
and the tagging mechanism for releases. As we don’t rely on a complex
branching strategy, the `branch` tags are unnecessary for our use case.